flask-cms-embed-tweet/requirements.txt
buttercms-python==1.3
Flask==2.0.3
tweepy==4.6.0